requirements.txt
Flask==3.1.1
flask-cors==6.0.1
Werkzeug==3.1.3
Jinja2==3.1.6
python-dotenv==1.1.1
requests==2.32.4
deepface==0.0.93
opencv-python==4.11.0.86
Pillow==11.3.0
pandas==2.3.0
matplotlib==3.10.3
threadpoolctl==3.6.0
scikit-learn==1.7.0
tensorflow==2.19.0
numpy==2.1.3
gunicorn==23.0.0
google-generativeai==0.8.5
